Lower power consumption during rendering
When running on HarmonyOS, the GPU will be used to render the background
if the background color of XComponent is any color other than
COLOR.Black, and the usage of GPU will cause high power consumption.
So enable the user to lower the power consumption by setting the background
color to COLOR.Black.

To address the black(or other color set by usr) flashing frame when
switching between ArkUI and Flutter pages, the background color should be
kept transparent until the onFirstFrame is called where changing the color
to the user's desired background color.

Signed-off-by: 小黑和平 <zhaohang14@huawei.com>