Modify message loop impl of ohos
It is need to read the timer_fd before the next cycle to call
epoll_wait, otherwist the epoll_wait will return immediately with the
same timer_fd event.
We should not rely on the main thread to read timer_fd in asynchronous
callback, because it will cause meaningless multiple calls to epoll_wait
between TimerFdWatcher uv_async_send and the asynchronous callback.

Signed-off-by: 小黑和平 <zhaohang14@huawei.com>